chore: rename theme-utils to theme
