+++ app/components/navbar.tsx
@@ -19,7 +19,7 @@ import {useEffect} from 'react'
 import {kodyProfiles} from '~/images.tsx'
 import {type OptionalTeam} from '~/utils/misc.tsx'
 import {useTeam} from '~/utils/team-provider.tsx'
-import {useOptimisticThemeMode} from '~/utils/theme-utils.tsx'
+import {useOptimisticThemeMode} from '~/utils/theme.tsx'
 import {useOptionalUser, useRootData} from '~/utils/use-root-data.ts'
 import {useElementState} from './hooks/use-element-state.tsx'
 import {LaptopIcon, MoonIcon, SunIcon} from './icons.tsx'
