+++ app/utils/mdx.tsx
@@ -26,7 +26,7 @@ import {type RootLoaderType} from '~/root.tsx'
 import {cache, cachified} from './cache.server.ts'
 import {markdownToHtmlUnwrapped, stripHtml} from './markdown.server.ts'
 import {getSocialMetas} from './seo.ts'
-import {Themed} from './theme-utils.tsx'
+import {Themed} from './theme.tsx'
 import {type Timings} from './timing.server.ts'
 import {useOptionalUser} from './use-root-data.ts'
 import {type MetaFunction, type TypedResponse} from '@remix-run/node'
